Extract larger route handlers into their own files.
