+++ oauth-proxy/index.js
@@ -16,7 +16,7 @@ const promBundle = require('express-prom-bundle');
 const Sentry = require('@sentry/node');
 const { logger, middlewareLogFormat } = require('./logger');
 
-const { loginBegin, loginEnd } = require('./metrics');
+const oauthHandlers = require('./oauthHandlers');
 
 const appRoutes = {
   authorize: '/authorization',
@@ -162,159 +162,15 @@ function buildApp(config, issuer, oktaClient, dynamo, dynamoClient) {
   });
 
   router.get(appRoutes.redirect, async (req, res, next) => {
-    const { state } = req.query;
-    if (!req.query.hasOwnProperty('error')) {
-      try {
-        await dynamoClient.saveToDynamo(dynamo, state, "code", req.query.code);
-      } catch (error) {
-        logger.error(`Failed to save authorization code in ${appRoutes.redirect} handler`, error);
-      }
-    }
-    try {
-      const document = await dynamoClient.getFromDynamoByState(dynamo, state);
-      const params = new URLSearchParams(req.query);
-      loginEnd.inc();
-      res.redirect(`${document.redirect_uri.S}?${params.toString()}`)
-    } catch (error) {
-      logger.error("Failed to redirect to the OAuth client application", error);
-      return next(error); // This error is unrecoverable because we can't look up the original redirect.
-    }
+    await oauthHandlers.redirectHandler(logger, dynamo, dynamoClient, req, res, next);
   });
 
   router.get(appRoutes.authorize, async (req, res, next) => {
-    loginBegin.inc();
-    const { state, client_id, redirect_uri: client_redirect } = req.query;
-    try {
-      const oktaApp = await oktaClient.getApplication(client_id);
-      if (oktaApp.settings.oauthClient.redirect_uris.indexOf(client_redirect) === -1) {
-        const errorParams = new URLSearchParams({
-          error: 'invalid_client',
-          error_description: 'The redirect URI specified by the application does not match any of the ' +
-                             `registered redirect URIs. Erroneous redirect URI: ${client_redirect}`
-        });
-        return res.redirect(`${client_redirect}?${errorParams.toString()}`);
-      }
-    } catch (error) {
-      logger.error("Unrecoverable error: could not get the Okta client app", error);
-      // This error is unrecoverable because we would be unable to verify
-      // that we are redirecting to a whitelisted client url
-      return next(error);
-    }
-
-    try {
-      await dynamoClient.saveToDynamo(dynamo, state, "redirect_uri", client_redirect);
-    } catch (error) {
-      logger.error(`Failed to save client redirect URI ${client_redirect} in ${appRoutes.authorize} handler`, error);
-      return next(error); // This error is unrecoverable because we can't create a record to lookup the requested redirect
-    }
-    const params = new URLSearchParams(req.query);
-    params.set('redirect_uri', redirect_uri);
-    if (!params.has('idp') && config.idp) {
-      params.set('idp', config.idp);
-    }
-    res.redirect(`${issuer.metadata.authorization_endpoint}?${params.toString()}`)
+    await oauthHandlers.authorizeHandler(config, redirect_uri, logger, issuer, dynamo, dynamoClient, oktaClient, req, res, next)
   });
 
   router.post(appRoutes.token, async (req, res, next) => {
-    let client_id, client_secret;
-
-    if (req.headers.authorization) {
-      const authHeader = req.headers.authorization.match(/^Basic\s(.*)$/);
-      ([ client_id, client_secret ] = Buffer.from(authHeader[1], 'base64').toString('utf-8').split(':'));
-    } else if (req.body.client_id && req.body.client_secret) {
-      ({ client_id, client_secret } = req.body);
-      delete req.body.client_id;
-      delete req.body.client_secret;
-    } else {
-      return res.status(401).json({
-        error: "invalid_client",
-        error_description: "Client authentication failed",
-      });
-    }
-
-    const client = new issuer.Client({
-      client_id,
-      client_secret,
-      redirect_uris: [
-        redirect_uri
-      ],
-    });
-
-    let tokens, state;
-    if (req.body.grant_type === 'refresh_token') {
-      try {
-        tokens = await client.refresh(req.body.refresh_token);
-      } catch (error) {
-        logger.error("Could not refresh the client session with the provided refresh token", error);
-        const statusCode = statusCodeFromError(error);
-        res.status(statusCode).json({
-          error: error.error,
-          error_description: error.error_description,
-        });
-      }
-      try {
-        const document = await dynamoClient.getFromDynamoBySecondary(dynamo, 'refresh_token', req.body.refresh_token);
-        state = document.state.S;
-        await dynamoClient.saveToDynamo(dynamo, state, 'refresh_token', tokens.refresh_token);
-      } catch (error) {
-        logger.error("Could not update the refresh token in DynamoDB", error);
-        state = null;
-      }
-    } else if (req.body.grant_type === 'authorization_code') {
-      try {
-        tokens = await client.grant(
-          {...req.body, redirect_uri }
-        );
-      } catch (error) {
-        logger.error("Failed to retrieve tokens using the OpenID client", error);
-        const statusCode = statusCodeFromError(error);
-        res.status(statusCode).json({
-          error: error.error,
-          error_description: error.error_description,
-        });
-      }
-      try {
-        const document = await dynamoClient.getFromDynamoBySecondary(dynamo, 'code', req.body.code);
-        state = document.state.S;
-        if (tokens.refresh_token) {
-          await dynamoClient.saveToDynamo(dynamo, state, 'refresh_token', tokens.refresh_token);
-        }
-      } catch (error) {
-        logger.error("Failed to save the new refresh token to DynamoDB", error);
-        state = null;
-      }
-    } else {
-      return res.status(400).json({
-        error: "unsupported_grant_type",
-        error_description: "Only authorization and refresh_token grant types are supported",
-      });
-    }
-
-    var decoded = jwtDecode(tokens.access_token);
-    if ((decoded.scp != null) && (decoded.scp.indexOf('launch/patient') > -1)) {
-      try {
-        const response = await requestPromise({
-          method: 'GET',
-          uri: config.validate_endpoint,
-          json: true,
-          headers: {
-            apiKey: config.validate_apiKey,
-            authorization: `Bearer ${tokens.access_token}`,
-          }
-        });
-        const patient = response.data.attributes.va_identifiers.icn;
-        res.json({...tokens, patient, state});
-      } catch (err) {
-        logger.error("Could not find a valid patient identifier for the provided authorization code", err);
-
-        res.status(400).json({
-          error: "invalid_grant",
-          error_description: "We were unable to find a valid patient identifier for the provided authorization code.",
-        });
-      }
-    } else {
-      res.json({...tokens, state});
-    }
+    await oauthHandlers.tokenHandler(config, redirect_uri, logger, issuer, dynamo, dynamoClient, req, res, next)
   });
 
   app.use(well_known_base_path, router)
