+++ oauth-proxy/oauthHandlers/index.js
@@ -1,0 +1,7 @@
+// This module should eventually contain an express route handler for each
+// official OAuth 2.0 endpoint.
+module.exports = {
+  'authorizeHandler': require('./authorizeHandler'),
+  'tokenHandler': require('./tokenHandler'),
+  'redirectHandler': require('./redirectHandler'),
+};
